+++ agent-control/tests/on_host/scenarios/filesystem_ops.rs
@@ -40,6 +40,7 @@ name: test
 version: 0.0.0
 platform: host
 operating_system: {AGENT_CONTROL_MODE_ON_HOST}
+protocol_version: "1.0"
 variables: {{}}
 deployment:
   filesystem:
@@ -141,6 +142,7 @@ name: test
 version: 0.0.0
 platform: host
 operating_system: {AGENT_CONTROL_MODE_ON_HOST}
+protocol_version: "1.0"
 variables:
   yaml_file_contents:
     description: "Contents of the YAML file"
@@ -309,6 +311,7 @@ name: infra-agent
 version: 0.0.0
 platform: host
 operating_system: {AGENT_CONTROL_MODE_ON_HOST}
+protocol_version: "1.0"
 variables:
   config_agent:
     description: "Agent configuration"